Refactor checklist and groupedChecklist into own files
